+++ packages/terra-avatar/src/terra-dev-site/test/avatar/avatar/ImageAvatarTallAndNarrow.test.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import Avatar from '../../../../index';
+import exampleAvatarImage from '../../../assets/170x251.jpg';
+
+export default () => <Avatar image={exampleAvatarImage} alt="User" id="image-avatar" />;