+++ packages/terra-form-field/src/Field.jsx
@@ -159,9 +159,10 @@ const Field = (props) => {
   );
 
   const content = React.Children.map(children, (child) => {
-    if (required && child && (child.type.isInput || child.type.isSelect || child.type.isTextarea)) {
+    if ((required || isInvalid) && child && (child.type.isInput || child.type.isSelect || child.type.isTextarea)) {
       return React.cloneElement(child, {
-        required: true,
+        ...required && { required: true },
+        ...isInvalid && { isInvalid: true },
       });
     }
     return child;