+++ packages/terra-form-field/tests/jest/Field.test.jsx
@@ -225,7 +225,7 @@ it('should render a required field with an Input', () => {
   expect(wrapper).toMatchSnapshot();
 });
 
-it('should render a required field with an Select', () => {
+it('should render a required field with a Select', () => {
   const field = (
     <Field
       label="Field Label"
@@ -239,7 +239,7 @@ it('should render a required field with an Select', () => {
   expect(wrapper).toMatchSnapshot();
 });
 
-it('should render a required field with an Textarea', () => {
+it('should render a required field with a Textarea', () => {
   const field = (
     <Field
       label="Field Label"
@@ -250,6 +250,93 @@ it('should render a required field with an Textarea', () => {
     </Field>
   );
   const wrapper = mountWithIntl(field).children();
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an invalid field with an Input', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      htmlFor="test"
+      isInvalid
+    >
+      <Input />
+    </Field>
+  );
+  const wrapper = mountWithIntl(field);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an invalid field with a Select', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      htmlFor="test"
+      isInvalid
+    >
+      <Select />
+    </Field>
+  );
+  const wrapper = mountWithIntl(field);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an invalid field with an Textarea', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      htmlFor="test"
+      isInvalid
+    >
+      <Textarea />
+    </Field>
+  );
+  const wrapper = mountWithIntl(field);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an invalid and required field with an Input', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      htmlFor="test"
+      isInvalid
+      required
+    >
+      <Input />
+    </Field>
+  );
+  const wrapper = mountWithIntl(field);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an invalid and required field with a Select', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      htmlFor="test"
+      isInvalid
+      required
+    >
+      <Select />
+    </Field>
+  );
+  const wrapper = mountWithIntl(field);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an invalid and required field with a Textarea', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      htmlFor="test"
+      isInvalid
+      required
+    >
+      <Textarea />
+    </Field>
+  );
+  const wrapper = mountWithIntl(field);
   expect(wrapper).toMatchSnapshot();
 });
 
@@ -266,7 +353,7 @@ it('should render a field with an Input', () => {
   expect(wrapper).toMatchSnapshot();
 });
 
-it('should render a field with an Select', () => {
+it('should render a field with a Select', () => {
   const field = (
     <Field
       label="Field Label"
@@ -279,7 +366,7 @@ it('should render a field with an Select', () => {
   expect(wrapper).toMatchSnapshot();
 });
 
-it('should render a field with an Textarea', () => {
+it('should render a field with a Textarea', () => {
   const field = (
     <Field
       label="Field Label"