+++ packages/terra-icon/scripts/src/migrate-svg/readSvg.js
@@ -2,7 +2,7 @@
 const fs = require('fs');
 
 const readSvg = csvObject => new Promise((resolve, reject) => {
-  const objectCsv = Object.assign({}, csvObject);
+  const objectCsv = { ...csvObject };
   fs.readFile(objectCsv.svgSrc, 'utf-8', (error, svg) => {
     if (error) {
       reject(error);