+++ packages/terra-button-group/CHANGELOG.md
@@ -7,9 +7,10 @@ Unreleased
 * Inline SVG aligned for IE and wdio screenshots Updated.
 * Import for package version from package.json added to doc page
 
-Unreleased
+3.21.0 - (August 21, 2019)
 ------------------
 ### Changed
+* Import for package version from package.json added to doc page
 * Cleaned up imports in examples and test files
 
 3.20.0 - (August 14, 2019)